% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmc.R
\name{cholesky}
\alias{cholesky}
\title{Manual Cholesky decomposition (lower-triangular)}
\usage{
cholesky(cov_mat)
}
\arguments{
\item{cov_mat}{Symmetric positive definite covariance matrix.}
}
\value{
Lower-triangular matrix \code{L}.
}
\description{
Computes the lower-triangular Cholesky factor \eqn{L} such that
\eqn{\Sigma = LL^\top}. Assumes input is symmetric positive definite.
}
\examples{
S <- matrix(c(4, 2, 2, 3), nrow = 2)
L <- cholesky(S)
all.equal(S, L \%*\% t(L))

}
