% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmc.R
\name{log_to_simple_returns}
\alias{log_to_simple_returns}
\title{Convert log-returns to simple returns}
\usage{
log_to_simple_returns(sim_log)
}
\arguments{
\item{sim_returns}{Matrix (or vector) of log-returns.}
}
\value{
Matrix (or vector) of simple log \code{exp(r) - 1}.
}
\description{
Convert log-returns to simple returns
}
