% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmc.R
\name{portfolio_pnl_from_returns}
\alias{portfolio_pnl_from_returns}
\title{Compute portfolio P&L from simulated returns}
\usage{
portfolio_pnl_from_returns(sim_simple_returns, weights, capital = 1e+06)
}
\arguments{
\item{weights}{Numeric vector of portfolio weights (length n assets).}

\item{capital}{Initial capital (default 1e6).}
}
\value{
Numeric vector of P&L (length = num_sims).
}
\description{
Takes simulated simple returns and converts them into portfolio P&L given weights and capital.
}
