% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varmc.R
\name{shock_gen}
\alias{shock_gen}
\title{Generate correlated shocks from a Cholesky factor}
\usage{
shock_gen(L, num_sims)
}
\arguments{
\item{L}{Lower-triangular Cholesky factor (from \code{cholesky()}).}

\item{num_sims}{Number of simulations (columns) to generate.}
}
\value{
Matrix of shocks with dimension (n assets) x (num_sims).
}
\description{
Draws i.i.d. standard normals and maps them through \eqn{L} to create correlated shocks.
}
\examples{
L <- diag(2)
shocks <- shock_gen(L, 5)
dim(shocks)

}
